requirements.txt
# Tooth Numbering YOLO System Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
PyYAML>=6.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional dependencies for enhanced functionality
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0